requirements-py3.txt
requests
protobuf>=3.4.0
six
lz4a